# [5.7] Optimize Collection::mapWithKeys