--- a/src/Illuminate/Support/Collection.php
+++ b/src/Illuminate/Support/Collection.php
@@ -1016,11 +1016,7 @@ class Collection implements ArrayAccess, Arrayable, Countable, IteratorAggregate
         $result = [];
 
         foreach ($this->items as $key => $value) {
-            $assoc = $callback($value, $key);
-
-            foreach ($assoc as $mapKey => $mapValue) {
-                $result[$mapKey] = $mapValue;
-            }
+            $result += $callback($value, $key);
         }
 
         return new static($result);