--- a/src/Illuminate/Support/Collection.php
+++ b/src/Illuminate/Support/Collection.php
@@ -1013,6 +1013,10 @@ class Collection implements ArrayAccess, Arrayable, Countable, IteratorAggregate
      */
     public function mapWithKeys(callable $callback)
     {
+        if ($this->isEmpty()) {
+            return new static([]);
+        }
+
         $values = array_map($callback, $this->items, array_keys($this->items));
 
         return new static(array_replace(...$values));