Benchmark (100,000 iterations):
```
Old method: 4.133s
New method: 3.175s
---
Old method: 3.754s
New method: 2.975s
---
Old method: 3.773s
New method: 3.246s
```

https://gist.github.com/dmitrybubyakin/6bdcc7c0ee7eaffdae699d61017214ee

15-25% faster.