@derekmd 2018-06-21
This would change the behavior from a merge to an append. Right now it's possible for a later `mapWithKeys()` iteration to overwrite the key of a previous iteration. This would make the first keyed value to always be used.