@vlakoff 2018-06-22
Beware of numeric keys, and order of overwriting.

Working solutions might be:
```php
$result = $callback($value, $key) + $result;
$result = array_replace($result, $callback($value, $key));
```

Note `array_replace()` seems to just be the same as `+` with reversed overwriting order, but probably significantly slower (function call overhead).

So, on the principle, why not changing to the `+` operator, if it's properly benchmarked and tested.