@vlakoff 2018-06-22
More importantly: https://github.com/laravel/framework/pull/16552#issuecomment-263143006

> Also take care of preserving the array order (not the same as the numerical indexes). To verify this, a foreach on the result should iterates the items in the same order.

That would be quite a subtle BC break... I haven't checked the proposal here about this. And tests should be added for this, if there are not yet.