@vlakoff 2018-06-22
The added test seems to be worth it, but I guess your new code is slower than current code, isn't it?