@dmitrybubyakin 2018-06-22
@vlakoff @tillkruss Added tests for overwriting keys. 

Benchmark (100,000 iterations):
```
Old method: 3.782s
New method: 2.757s
---
Old method: 4.678s
New method: 3.075s
---
Old method: 4.173s
New method: 3.014s
```

28 - 35% now.