@dmitrybubyakin 2018-06-22
@vlakoff I made some changes, so my new code is faster than the previous code. But there is more memory usage.