@vlakoff 2018-06-22
I see about memory usage. This should be measured as well. With your code here, we get a *huge* `array_replace()` call with many parameters!

Have you tried my simple `$result = $callback($value, $key) + $result;`?