@sisve 2018-06-22
While the current implementation says "The callback should return an associative array with a single key/value pair", it supports a callback returning multiple entries. Is this something that is still supported, or are we dropping that support?