@dmitrybubyakin 2018-06-22
@vlakoff Yes, I have. `$result = $callback($value, $key) + $result;` is much slower than the old code.

@sisve We are nothing dropping.