@dmitrybubyakin 2018-06-22
@vlakoff 

```
10000 iterations over 100 items, strlen 10:
Old method: time 371ms, memory: 2.00Mb
New method: time 307ms, memory: 2.00Mb
----
10000 iterations over 100 items, strlen 100:
Old method: time 382ms, memory: 2.00Mb
New method: time 294ms, memory: 2.00Mb
----
10000 iterations over 100 items, strlen 1000:
Old method: time 376ms, memory: 2.00Mb
New method: time 298ms, memory: 2.00Mb
----

10000 iterations over 1000 items, strlen 100:
Old method: time 3.682s, memory: 2.00Mb
New method: time 2.987s, memory: 4.00Mb
----
100 iterations over 10000 items, strlen 1000:
Old method: time 603ms, memory: 22.00Mb
New method: time 597ms, memory: 26.00Mb
----
10 iterations over 100000 items, strlen 1000:
Old method: time 1.065s, memory: 181.50Mb
New method: time 988ms, memory: 221.51Mb
----
```