@vlakoff 2018-06-22
I tried to reproduce the performance increase, but on my machine your code is actually a bit slower...

```php
$nb = 100;

$data = range(1, 1000);

$callback = function ($value, $key) {
    return [$key => $value];
};


$t1 = microtime(true);
for ($ii = $nb; $ii--; ) {

    $result = [];
    foreach ($data as $key => $value) {
        $assoc = $callback($value, $key);
        foreach ($assoc as $mapKey => $mapValue) {
            $result[$mapKey] = $mapValue;
        }
    }

}
$t2 = microtime(true);
for ($ii = $nb; $ii--; ) {

    $values = array_map($callback, $data, array_keys($data));
    $result = array_replace(...$values);

}
$t3 = microtime(true);

echo $t2 - $t1;
echo "\n";
echo $t3 - $t2;

echo "\n\n";
echo $t2 - $t1 == 0 ? 'N/A' : ($t3 - $t2) * 100 / ($t2 - $t1);
```

Both methods seem to be O(n).