@dmitrybubyakin 2018-06-22
@vlakoff on my machine (php 7.2.6):

```
0.01982307434082
0.012885093688965

65.000481093043%
```