@vlakoff 2018-06-22
Yeah, I had tested on a old PHP 5 box. On PHP 7, both codes are much faster, and yours seems to be faster indeed.