@vlakoff 2018-06-22
Again on my box, above a certain amount of items, your code gets suddenly slower. I guess there is some memory exhaustion, which breaks scalability.