@dmitrybubyakin 2018-06-22
@vlakoff 

https://3v4l.org/YuZn8/perf#perfomance
https://3v4l.org/kone1/perf#perfomance