@vlakoff 2018-06-22
I can confirm on 3v4l.org that on large data (try 10000, 50000...), your code performs slower and uses more memory. There seems to be a threshold at which execution speed seems to fall down, and memory usage difference seems to gradually increase with data size.

For me scalability is the most important by far, even if it means slower (yet acceptable) execution speed of smaller datasets. Memory usage shouldn't be neglected too.