@vlakoff 2018-06-22
Just as a reminder, some possible subtle BC breaks have been highlighted on this thread, and unit tests may be added for these.