@vlakoff 2018-06-24
@dmitrybubyakin I suggest you open a new PR, for just adding this unit test concerning numeric keys overwriting.

The other cases I mentioned (callback returning multiple rows, preservation of array order) are actually already covered by tests.